refactor(firestore): extract per-family modules behind the entry barrel
Move every inline operation out of src/firestore/index.ts into per-family
modules and reduce the entry to a pure re-export barrel. Mechanical moves
only — bodies are verbatim; the sole edit is the import header each module
needs plus the barrel re-exports.

Families (each well under the 600-line split threshold), mirroring the
symmetric pyric-tools worker client split:
  instances, equality, persistence, refs, reads, writes,
  query-constraints, aggregates, listeners, transactions,
  field-values, sandbox-ops.

The barrel re-exports the public surface exactly: the frozen runtime
export set and the full typed export surface are byte-identical before and
after. Package-internal state helpers stay off the published subpath (only
TARGET_SYMBOL is re-exported from state).
